Refactor BaseFormula class to include string representation and analytical formula
This commit refactors the BaseFormula class in the _utility/base_formula.py module. It adds a __str__ and __repr__ method to provide a string representation of the formula. It also includes a _get_str method to return the analytical formula as a string. These changes improve the readability and usability of the BaseFormula class.

Refactor functions.py module and add new distance classes

This commit refactors the functions.py module and introduces new distance classes: AngularSeparation, LosDistance, RadialDistance2D, RadialDistance3D, TotalVelocity, and EffectivePotential. These classes provide formulas and methods for computing various distances and velocities. The refactoring improves the organization and modularity of the code.

Changed function module intelligence, modified inits, fixed bugs

This commit includes changes to the function module. It improves the intelligence of the module by modifying the initializations and fixing bugs. These changes enhance the functionality and reliability of the function module.

Implemented R code interaction in plots for regression

This commit implements R code interaction in plots for regression. It enhances the regression plots by incorporating R code, improving the visualization and analysis capabilities.
It also gratly modifies functions logics and computations, especially for analytical and numerical errors computation.

Little bug fix

This commit includes a small bug fix. It addresses a minor issue in the code, improving the overall stability and correctness.