Refactor compute_error function in calculus.py to handle correlations
The compute_error function in the calculus.py module has been refactored to handle correlations between variables. This allows for more accurate error propagation in calculations. The function now takes an additional parameter, corr_values, which is a list of correlation values. If corr_values is provided, the function includes the correlations in the computation of the error.

This change improves the accuracy of error calculations in the ggcas package.